Set encoding to UTF8 without BOM
`Encoding.UTF8` will include a UTF8 BOM character, which is invalid according to the [JSON spec RFC7159, Section 8.1](https://tools.ietf.org/html/rfc7159#section-8.1).
This was causing issues with (in particular) JavaScript clients that call `JSON.parse` on the response.
`UTF8Encoding` allows us to specify whether to include the BOM character or not.